"Not found" handling.

+++ auth/models/user.py
@@ -95,7 +95,10 @@ class User(db.Model, ModelMixin):
     @classmethod
     def delete(cls, user_id: UUID) -> bool:
         user = cls.query.filter_by(id=user_id).first()
+        if user is None:
+            return False
         result = db.session.delete(user)
+        logger.info('deleting result: %s', result)
         db.session.commit()
         logger.info('User deleted: %s', user_id)
         return True
